% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ids.R
\name{count_ids}
\alias{count_ids}
\title{Count unique household and individual observations}
\usage{
count_ids(data, groupvars = NULL, hhid = hhid, pid = pid)
}
\arguments{
\item{data}{a data frame}

\item{groupvars}{a quoted vector of variables that define uniqueness across surveys}

\item{hhid}{the household id variable}

\item{pid}{the individual id variable}
}
\value{
a tibble
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{cound_ids} provides and output data.frame of the number of unique
household ids and individual ids within a user-defined grouping
}
\examples{
\dontrun{
# Case where survey variables hhid and pid determine uniqueness
yearly_data \%>\%
  count_ids(hhid, pid)

# Case where survey observations are unique by hhid and pid within each year
# and are now appended together in multi-year tibble
appended_yearly_data \%>\%
  count_ids(groupvars = c(year), hhid, pid)
}
}
