% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_surveys.R
\name{import_surveys}
\alias{import_surveys}
\title{Import multiple surveys}
\usage{
import_surveys(
  eval_directory,
  filetype = "dta",
  vars = NULL,
  version = NULL,
  file_pattern = paste0("\\\\", version, "_ALL.dta$")
)
}
\arguments{
\item{eval_directory}{the top level directory folder}

\item{filetype}{one of "dta", "rda", or "rds".}

\item{vars}{additional quoted variables to be included beyond standard survey
identifiers}

\item{version}{a quoted expression that indicates the file version to import}

\item{file_pattern}{a quoted regex expression to match file patterns in the
directory}
}
\value{
a dataframe of all appended survey files
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{import_surveys} imports all survey files in a given directory and
outputs an appended dataframe
}
\examples{
\dontrun{
import_surveys("/root/docs/surveys", filetype = "rda", file_pattern = ".") # all files in directory
import_surveys("/root/docs/surveys", filetype = "rda", file_pattern = "annual_survey")
import_surveys("/root/docs/surveys",
                filetype = "rda",
                vars = c("Var1", "Var3"),
                file_pattern = "annual_survey")
import_surveys("/root/docs/surveys",
                filetype = "dta",
                version = "V_03",
                file_pattern = paste0("\\\\", version, "_survey.dta$"))
}
}
