% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_codes.R
\name{match_codes}
\alias{match_codes}
\title{Determine most likely matches between two vectors.}
\usage{
match_codes(
  df,
  country_code,
  international_code,
  pad_vars = NULL,
  check_matches = FALSE,
  default_grepl = "[A-Za-z]",
  match_bins = 5
)
}
\arguments{
\item{df}{The tibble or data.frame object containing the two columns of data
to match}

\item{country_code}{The name of the column that contains the country-specific
international_code}

\item{international_code}{The name of the column that contains the
international code}

\item{pad_vars}{A quoted character vector of variables to pad, otherwise NULL}

\item{check_matches}{Run checks to verify quality of results, TRUE or FALSE}

\item{default_grepl}{remove rows that do not match this pattern}

\item{match_bins}{the number of bins in the density plot for match quality}
}
\value{
a 4-element list object that contains the final match tibble, a
results tibble, a ggplot by match quality and a ggplot by match stage.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{match_codes} creates a one-to-one key between two classification
schemes. Given two vectors, it first finds finding unique one-to-one matches,
and then systematically determines most likely matches for imperfect original
correspondences.
}
\examples{
\dontrun{
# works with any set of two string vectors
starwars \%>\%
  match_codes(
      country_code = homeworld,
      international_code = eye_color,
      pad_vars = NULL,
      check_matches = FALSE,
      default_grepl = "[:digit:]",
      match_bins = 20)
 }
}
