% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdf.R
\name{read_pdf}
\alias{read_pdf}
\title{Read information from a table in PDF format}
\usage{
read_pdf(
  pdf_path,
  page_min,
  page_max,
  header = TRUE,
  varnames = c("var1", "var2", "var3", "var4", "var5"),
  ymin = 90,
  xlabel = c(155, 420),
  xmin = c(91, 131, 415, 446, 501),
  xmax = c(130, 175, 445, 500, 9999),
  numlist = NULL,
  fuzzy_rows = FALSE,
  match_tol = 2
)
}
\arguments{
\item{pdf_path}{The file path to the pdf to read}

\item{page_min}{The starting page number (inclusive) to begin reading the file.}

\item{page_max}{The ending page number (inclusive) to end reading the file.}

\item{header}{A boolean statement indicating if the document always has a constant page header on all pages}

\item{varnames}{A list of variable names of the output columns}

\item{ymin}{The minimum y coordinate on the page where the function will read information}

\item{xlabel}{a 2-element list that indicates start and ending x-coordinates of omitted columns}

\item{xmin}{a list of x-coordinate column minimums. Same length as varnames.}

\item{xmax}{a list of x-coordinate column maximums. Same length as varnames.}

\item{numlist}{an optional vector of omitted numbers to filter.}

\item{fuzzy_rows}{an optional boolean parameter that toggles slightly misaligned rows in pdf}

\item{match_tol}{the nearest-neighbor y match tolerance for misaligned rows}
}
\value{
a tibble with tables read from a pdf file
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{read_pdf} extracts table multipage information into a tibble. This is a
very tricky operation that largely depends on trial and error of column
parameter arguments.
}
\examples{
\dontrun{
read_pdf(
        header = TRUE,
        varnames = c("ingredient", "grams", "ounces"),
        ymin = 50,
        xlabel = c(150,400),
        xmin = c(0,400,500),
        xmax = c(100,500,1000),
        numlist = NULL,
        fuzzy_rows = FALSE
        )

read_pdf(
        header = FALSE,
        varnames = c("var1", "var2")),
        ymin = 50,
        xlabel = c(500,1000),
        xmin = c(0,90,200),
        xmax = c(100,200,1000),
        numlist = NULL,
        fuzzy_rows = TRUE,
        match_tol = 3
        )
}
}
