% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_pdf_cols.R
\name{reveal_pdf_cols}
\alias{reveal_pdf_cols}
\title{Examine column-ordered data in a pdf}
\usage{
reveal_pdf_cols(pdf_path, page_min, page_max, nbins = 100)
}
\arguments{
\item{pdf_path}{the quoted path to the pdf file to read}

\item{page_min}{the page to begin reading, inclusive}

\item{page_max}{the page to end reading, inclusive}

\item{nbins}{the number of histogram bins}
}
\value{
a histogram graph
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{reveal_pdf_cols} creates a histogram of elements of x-axis positions of
the pdf to read. The pdf is read in a semi-processed form and the function
then creates a light interactive output, designed to help the user prepare the
column parameters for \code{read_pdf}.
}
\examples{
\dontrun{
reveal_pdf_cols("/root/docs/my_cooking_pdf.pdf", 1, 3)
pdf_hist <- reveal_pdf_cols("/root/docs/my_cooking_pdf.pdf", 1, 3, nbins = 50)
}
}
